Change numEntries to a double to hold fractional fills.
--HG--
branch : multiweight

+++ include/YODA/Bin2D.h
@@ -289,7 +289,7 @@ namespace YODA {
     }
 
     /// The number of entries
-    unsigned long numEntries() const {
+    double numEntries() const {
       return _dbn.numEntries();
     }
