+++ include/YODA/Dbn3D.h
@@ -28,7 +28,7 @@ namespace YODA {
     /// Constructor to set a distribution with a pre-filled state.
     ///
     /// Principally designed for internal persistency use.
-    Dbn3D(unsigned long numEntries,
+    Dbn3D(double numEntries,
           double sumW, double sumW2,
           double sumWX, double sumWX2,
           double sumWY, double sumWY2,
@@ -231,7 +231,7 @@ namespace YODA {
     //@{
 
     /// Number of entries (number of times @c fill was called, ignoring weights)
-    unsigned long numEntries() const {
+    double numEntries() const {
       return _dbnX.numEntries();
     }
