+++ include/YODA/Profile2D.h
@@ -295,8 +295,8 @@ namespace YODA {
     /// @name Whole histo data
     //@{
 
-    /// Get the number of fills
-    unsigned long numEntries(bool includeoverflows=true) const;
+    /// Get the number of fills (fractional fills are possible)
+    double numEntries(bool includeoverflows=true) const;
 
     /// Get the effective number of fills
     double effNumEntries(bool includeoverflows=true) const;
