+++ src/Profile2D.cc
@@ -49,7 +49,7 @@ namespace YODA {
 
   /////////////// COMMON TO ALL BINNED
 
-  unsigned long Profile2D::numEntries(bool includeoverflows) const {
+  double Profile2D::numEntries(bool includeoverflows) const {
     if (includeoverflows) return totalDbn().numEntries();
     unsigned long n = 0;
     for (const Bin& b : bins()) n += b.numEntries();
